Gets log entries logging to the database

+++ envcage-recorder/src/payload.rs
@@ -1,20 +1,21 @@
-use std::str::FromStr;
-use serde::Deserialize;
+use super::error::Error;
+use bigdecimal::BigDecimal;
 use chrono::prelude::*;
 use chrono::Utc;
-use super::error::Error;
+use serde::Deserialize;
+use std::str::FromStr;
 
 #[derive(Deserialize, Debug)]
 pub(crate) struct Payload {
-	pub created: DateTime<Utc>,
-	pub temperature: f32,
-	pub humidity: f32,
+    pub created: DateTime<Utc>,
+    pub temperature: BigDecimal,
+    pub humidity: BigDecimal,
 }
 
 impl FromStr for Payload {
-	fn from_str(s: &str) -> Result<Payload, Self::Err> {
-		serde_json::from_str(s).map_err(|e| Error::PayloadParse(format!("{}\n--\n{:?}", s, e)))
-	}
+    fn from_str(s: &str) -> Result<Payload, Self::Err> {
+        serde_json::from_str(s).map_err(|e| Error::PayloadParse(format!("{}\n--\n{:?}", s, e)))
+    }
 
-	type Err = Error;
+    type Err = Error;
 }
